Impl the projection, per Rohlf 1999 equation 1.
I throw a warning to the user if they've specified do_project alongside do_scaling. The 1999 paper assumes that all shapes are of unit size,
which makes sense if they are to be on a unit sphere. The reference X_c is actually not assumed to be of unit size, although a note suggests
that it is almost always the case.