Change license value to 'MIT' instead of the entire LICENSE file contents.
I had this idea after running `twine check dist/*` and trying to investigate the failure.
Apparently there was an indentation error on line 21 of my readme, which didn't make sense to me,
because the readme is not even 21 lines long. I then unpacked the wheel and looked at the
METADATA, which looked screwy, and had the entire license pasted on the top. Hence the idea.
Dunno if it will work tho.