Switch to opt 3 in 'Single-sourcing the package version' for versioning, for now
I had to do this because I was getting "ModuleNotFoundError: No module named
'version_helper'" on a clean machine.
The reason I was not getting this on my dev machine is because I was always in
the root dev folder. So "version_helper" was always found. Lucky me.

Still, what I'm using here is obviously a sub-optimal choice, because it
creates two files that need updating when the version has to change (the
morphops/_version.py, as well as the conf.py for sphinx).

I also don't like that setup.py has to know about the morphops directory, but I
don't think there is a way around that.
Lastly, there is no way to integrate this with VCS tags. Will have to try using
pbr or something like that.